requirements.txt
mongoengine
discord.py[voice]
websockets
aiohttp
azure

